requirements.txt
streamlit
python-dotenv
openai
tiktoken
langchain
langchain-community
langchain-openai
pydantic
